+++ src/twitch/handler/CameraSettingsPopup.cpp
@@ -6,7 +6,8 @@
 using namespace cocos2d;
 using namespace geode::prelude;
 
-bool CameraSettingsPopup::setup() {
+bool CameraSettingsPopup::setup()
+{
     setTitle("Camera Settings");
     setID("camera-settings-popup");
 
@@ -14,8 +15,8 @@ bool CameraSettingsPopup::setup() {
     auto m_contentLayer = CCLayer::create();
 
     m_contentLayer->setContentSize(popupSize);
-    m_contentLayer->setAnchorPoint({ 0, 0 });
-    m_contentLayer->setPosition({ 0, 0 });
+    m_contentLayer->setAnchorPoint({0, 0});
+    m_contentLayer->setPosition({0, 0});
 
     m_noElasticity = true;
     addChild(m_contentLayer);
@@ -33,26 +34,27 @@ bool CameraSettingsPopup::setup() {
     float inputY = rowY - 20.f;
 
     // Labels and inputs in a row: Skew, Rotation, Scale, Time
-    const char* names[] = { "Skew", "Rotation", "Scale", "Time" };
-    geode::TextInput** inputs[] = { &m_skewInput, &m_rotInput, &m_scaleInput, &m_timeInput };
+    const char *names[] = {"Skew", "Rotation", "Scale", "Time"};
+    geode::TextInput **inputs[] = {&m_skewInput, &m_rotInput, &m_scaleInput, &m_timeInput};
 
-    for (int i = 0; i < 4; ++i) {
+    for (int i = 0; i < 4; ++i)
+    {
         float x = rowStartX + i * (fieldWidth + spacing);
 
         // Label above input
         auto label = CCLabelBMFont::create(names[i], "bigFont.fnt");
         label->setScale(0.5f);
-        label->setAnchorPoint({ 0.5f, 0.5f });
+        label->setAnchorPoint({0.5f, 0.5f});
         label->setPosition(x + fieldWidth / 2, labelY);
 
         m_contentLayer->addChild(label);
 
-        // Input
-        *inputs[i] = geode::TextInput::create(fieldWidth, names[i], "bigFont.fnt");
+        // i found out that u can do this if you make an array cuz me lazy xdd
+        *inputs[i] = TextInput::create(fieldWidth, names[i], "bigFont.fnt");
 
         (*inputs[i])->setPosition(x + fieldWidth / 2, inputY);
         (*inputs[i])->setScale(0.6f);
-        (*inputs[i])->setMaxCharCount(6);
+        (*inputs[i])->setCommonFilter(CommonFilter::Float);
 
         m_contentLayer->addChild(*inputs[i]);
     };
@@ -72,8 +74,10 @@ bool CameraSettingsPopup::setup() {
     return true;
 };
 
-void CameraSettingsPopup::onSave(CCObject* sender) {
-    if (m_callback) {
+void CameraSettingsPopup::onSave(CCObject *sender)
+{
+    if (m_callback)
+    {
         float skew = getSkew();
         float rot = getRotation();
         float scale = getScale();
@@ -84,43 +88,53 @@ void CameraSettingsPopup::onSave(CCObject* sender) {
     onClose(sender);
 };
 
-std::string CameraSettingsPopup::formatCameraLabel(float skew, float rot, float scale, float time) {
+std::string CameraSettingsPopup::formatCameraLabel(float skew, float rot, float scale, float time)
+{
     return fmt::format("Skew: {:.2f}, Rot: {:.2f}, Scale: {:.2f}, Time: {:.2f}s", skew, rot, scale, time);
 };
 
-void CameraSettingsPopup::onApply(CCObject* sender) {
+void CameraSettingsPopup::onApply(CCObject *sender)
+{
     // Optionally apply changes live
     if (m_callback)
         m_callback(getSkew(), getRotation(), getScale(), getTime());
 };
 
-void CameraSettingsPopup::onClose(CCObject* sender) {
+void CameraSettingsPopup::onClose(CCObject *sender)
+{
     removeFromParentAndCleanup(true);
 };
 
-float CameraSettingsPopup::getSkew() const {
+float CameraSettingsPopup::getSkew() const
+{
     return numFromString<float>(m_skewInput->getString()).unwrapOrDefault();
 };
 
-float CameraSettingsPopup::getRotation() const {
+float CameraSettingsPopup::getRotation() const
+{
     return numFromString<float>(m_rotInput->getString()).unwrapOrDefault();
 };
 
-float CameraSettingsPopup::getScale() const {
+float CameraSettingsPopup::getScale() const
+{
     return numFromString<float>(m_scaleInput->getString()).unwrapOrDefault();
 };
 
-float CameraSettingsPopup::getTime() const {
+float CameraSettingsPopup::getTime() const
+{
     return numFromString<float>(m_timeInput->getString()).unwrapOrDefault();
 };
 
-CameraSettingsPopup* CameraSettingsPopup::create(float skew, float rot, float scale, float time, std::function<void(float, float, float, float)> callback) {
+CameraSettingsPopup *CameraSettingsPopup::create(float skew, float rot, float scale, float time, std::function<void(float, float, float, float)> callback)
+{
     auto ret = new CameraSettingsPopup();
 
-    if (ret) {
+    if (ret)
+    {
         ret->m_callback = callback;
 
-        if (ret->initAnchored(330.f, 180.f)) {
+        if (ret->initAnchored(330.f, 180.f))
+        {
             ret->autorelease();
 
             ret->m_skewInput->setString(fmt::format("{:.2f}", skew).c_str());