+++ changelog.md
@@ -4,6 +4,9 @@
 - Fixed the **Key Code** action node formatting
 - Added Close confirmation on the Command Settings and editing commands when escape button is clicked
 - Added extra options for the **Sound Effect** action settings (Speed, Volume, Pitch, Start, End)
+- Fixed **Wait** action time value disappearing when adding/moving actions.
+- Replaced the input fonts to **Pusab**
+
 # v0.1.10-beta.1
 - Added **Experimental Features** option in the Mod Settings.
 - Fixed Command Cooldown not detecting correctly.